+++ forge/src/main/java/de/cech12/ceramicshears/ForgeCeramicShearsMod.java
@@ -2,11 +2,12 @@ package de.cech12.ceramicshears;
 
 import de.cech12.ceramicshears.item.CeramicShearsItem;
 import net.minecraft.core.dispenser.ShearsDispenseItemBehavior;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.DispenserBlock;
 import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
-import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -14,6 +15,8 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
+
+import java.util.function.Function;
 
 /**
  * Mod class for the Forge loader.
@@ -27,20 +30,23 @@ public class ForgeCeramicShearsMod {
     public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, Constants.MOD_ID);
 
     /** clay shears part item registry object */
-    public static final RegistryObject<Item> CLAY_SHEARS_PART = ITEMS.register("clay_shears_part", () -> new Item(new Item.Properties()));
+    public static final RegistryObject<Item> CLAY_SHEARS_PART = registerItem("clay_shears_part", Item::new);
     /** ceramic shears part item registry object */
-    public static final RegistryObject<Item> CERAMIC_SHEARS_PART = ITEMS.register("ceramic_shears_part", () -> new Item(new Item.Properties()));
+    public static final RegistryObject<Item> CERAMIC_SHEARS_PART = registerItem("ceramic_shears_part", Item::new);
 
     static {
-        Constants.CERAMIC_SHEARS = ITEMS.register("ceramic_shears", CeramicShearsItem::new);
+        Constants.CERAMIC_SHEARS = registerItem("ceramic_shears", CeramicShearsItem::new);
+    }
+
+    private static RegistryObject<Item> registerItem(String name, Function<Item.Properties, Item> itemConstructor) {
+        return ITEMS.register(name, () -> itemConstructor.apply(new Item.Properties().setId(ResourceKey.create(BuiltInRegistries.ITEM.key(), Constants.id(name)))));
     }
 
     /**
      * Constructor of a mod instance.
      */
-    public ForgeCeramicShearsMod() {
-        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-        ITEMS.register(modEventBus);
+    public ForgeCeramicShearsMod(FMLJavaModLoadingContext context) {
+        ITEMS.register(context.getModEventBus());
         CommonLoader.init();
     }